ideas/daidokoro-note/ocr-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
paddleocr==2.9.1
paddlepaddle==3.0.0b1
pillow==11.0.0
python-dotenv==1.0.1
httpx==0.28.0
ulid-py==1.1.0
